requirements.txt
flask
flask-cors
flask-sqlalchemy
pyjwt
werkzeug